Refactor sign-up
